+++ obj/src/persistence/IdentifiablePostgresPersistence.d.ts
@@ -85,6 +85,10 @@ import { PostgresPersistence } from './PostgresPersistence';
  */
 export declare class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> extends PostgresPersistence<T> implements IWriter<T, K>, IGetter<T, K>, ISetter<T> {
     /**
+     * Flag to turn on automated string ID generation
+     */
+    protected _autoGenerateId: boolean;
+    /**
      * Creates a new instance of the persistence component.
      *
      * @param tableName    (optional) a table name.